# clawos/docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ClawOS Middleware API
  version: 0.2.0
  description: Contracts for agent orchestration, shared memory, voice processing, handoff, and project status.
  license:
    name: Proprietary
    url: https://example.com/legal/clawos-license
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /voice/process:
    post:
      operationId: processVoice
      summary: Process voice input and return transcript plus response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceProcessRequest"
      responses:
        "200":
          description: Voice pipeline output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceProcessResponse"
        "400":
          description: Invalid or missing voice payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Audio payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: External voice provider runtime failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Voice provider configuration missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /voice/output/{fileName}:
    get:
      operationId: getVoiceOutput
      summary: Stream persisted TTS output audio file
      parameters:
        - in: path
          name: fileName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audio output stream (mp3)
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Voice output file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/status:
    get:
      operationId: getProjectStatus
      summary: Consolidated project status
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
        - in: query
          name: probe
          required: false
          schema:
            type: boolean
          description: When true, middleware attempts live websocket probe to OpenClaw daemon.
      responses:
        "200":
          description: Project status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStatusResponse"
        "400":
          description: Invalid project query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/events:
    get:
      operationId: getProjectEvents
      summary: Recent thought and action events for one project
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Project runtime events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEventsResponse"
        "400":
          description: Invalid project query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /knowledge/feed:
    get:
      operationId: getKnowledgeFeed
      summary: Global discovery feed entries
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Feed entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeFeedResponse"
        "400":
          description: Missing project_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VoiceProcessRequest:
      type: object
      required:
        - audio_base64
        - agent_id
      properties:
        audio_base64:
          type: string
          description: Base64 encoded audio clip
        agent_id:
          type: string
        project_id:
          type: string
        voice_id:
          type: string
    VoiceProcessResponse:
      type: object
      properties:
        transcript:
          type: string
        agent_response:
          type: string
        tts_audio_url:
          type: string
        openclaw_routed:
          type: boolean
          description: True when transcript was delivered to OpenClaw websocket bridge.
    ProjectStatusResponse:
      type: object
      properties:
        project_id:
          type: string
        overall_status:
          type: string
        agents:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              status:
                type: string
        tasks:
          type: object
          additionalProperties:
            type: integer
        bridge:
          type: object
          properties:
            ws_url:
              type: string
            connected:
              type: boolean
        updated_at:
          type: string
          format: date-time
    ProjectEvent:
      type: object
      required:
        - id
        - project_id
        - kind
        - content
        - timestamp
      properties:
        id:
          type: string
        project_id:
          type: string
        agent_id:
          type: string
          nullable: true
        agent_name:
          type: string
          nullable: true
        kind:
          type: string
          enum: [thought, action]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    ProjectEventsResponse:
      type: object
      properties:
        project_id:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/ProjectEvent"
...
    KnowledgeFeedResponse:
      type: object
      properties:
        project_id:
          type: string
        entries:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              agent_id:
                type: string
              message:
                type: string
              timestamp:
                type: string
                format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true